style: Move BorderStyle to border.rs.
Differential Revision: https://phabricator.services.mozilla.com/D12858
